+++ borax/calendars/festivals2.py
@@ -8,7 +8,7 @@ from datetime import date, timedelta
 from pathlib import Path
 from typing import List, Tuple, Optional, Union, Iterator
 
-from borax.calendars.lunardate import LunarDate, LCalendars, TermUtils
+from borax.calendars.lunardate import LunarDate, LCalendars, TermUtils, TextUtils
 
 __all__ = [
     'FestivalError', 'WrappedDate', 'Period',
@@ -176,6 +176,7 @@ class Festival:
         self._freq = kwargs.get('freq', 0)
 
         self._cyear = 0
+        self._tmp = {}
 
     def gets(self, *args):
         def _get(_f):
@@ -201,6 +202,13 @@ class Festival:
     @property
     def name(self):
         return self._name
+
+    def _get_description(self) -> str:
+        pass
+
+    @property
+    def description(self) -> str:
+        return self._get_description()
 
     def is_(self, date_obj: MixedDate) -> bool:
         date_obj = self._normalize(date_obj)
@@ -361,6 +369,26 @@ class SolarFestival(Festival):
             reverse = 0
         super().__init__(name=name, freq=freq, month=month, day=day, reverse=reverse)
 
+    def _get_description(self) -> str:
+        cn_list = ['公历']
+        day_order = False
+        if self._freq == FreqConst.YEARLY:
+            if self._month != 0:
+                cn_list.append('每年{}月'.format(self._month))
+            else:
+                cn_list.append('每年')
+                day_order = True
+        else:
+            cn_list.append('每月')
+        if self._reverse:
+            cn_list.append('倒数第{}天'.format(self._day))
+        else:
+            if day_order:
+                cn_list.append('第{}天'.format(self._day))
+            else:
+                cn_list.append('{}日'.format(self._day))
+        return ''.join(cn_list)
+
     def _resolve_yearly(self, year) -> List[Union[date, LunarDate]]:
         if self._month == 0:
             if self._reverse == 0:
@@ -418,6 +446,9 @@ class WeekFestival(Festival):
     def __init__(self, *, month, index, week, name=None):
         super().__init__(name=name, freq=FreqConst.YEARLY, month=month, index=index, week=week)
 
+    def _get_description(self) -> str:
+        return '公历{}月第{}个星期{}'.format(self._month, self._week_index, '一二三四五六日'[self._week_no])
+
     def _resolve_yearly(self, year) -> List[Union[date, LunarDate]]:
         day = WeekFestival.week_day(year, self._month, self._week_index, self._week_no)
         return [date(year, self._month, day)]
@@ -444,7 +475,12 @@ class TermFestival(Festival):
     def __init__(self, *, index=None, name=None):
         if index is None:
             index = TermUtils.get_index_for_name(name)
+        else:
+            name = TermUtils.get_name_for_index(index)
         super().__init__(freq=FreqConst.YEARLY, name=name, index=index)
+
+    def _get_description(self) -> str:
+        return '公历每年{}节气'.format(self.name)
 
     def _resolve_yearly(self, year) -> List[Union[date, LunarDate]]:
         try:
@@ -470,6 +506,26 @@ class LunarFestival(Festival):
         else:
             reverse = 0
         super().__init__(freq=freq, name=name, month=month, day=day, leap=leap, reverse=reverse)
+
+    def _get_description(self) -> str:
+        cn_list = ['农历']
+        day_order = False
+        if self._freq == FreqConst.YEARLY:
+            if self._month != 0:
+                cn_list.append('每年{}月'.format(TextUtils.month_cn(self._month)))
+            else:
+                cn_list.append('每年')
+                day_order = True
+        else:
+            cn_list.append('每月')
+        if self._reverse:
+            cn_list.append('倒数第{}天'.format(self._day))
+        else:
+            if day_order:
+                cn_list.append('第{}天'.format(self._day))
+            else:
+                cn_list.append('{}'.format(TextUtils.day_cn(self._day)))
+        return ''.join(cn_list)
 
     def _resolve_yearly(self, year: int) -> List[Union[date, LunarDate]]:
         month_meta = list(LCalendars.iter_year_month(year))